+++ src/main/java/org/folio/entitlement/service/flow/EntitleFlowFactory.java
@@ -7,9 +7,6 @@ import lombok.RequiredArgsConstructor;
 import org.folio.entitlement.domain.model.EntitlementRequest;
 import org.folio.entitlement.integration.kafka.KafkaTenantTopicCreator;
 import org.folio.entitlement.service.stage.ApplicationDescriptorTreeLoader;
-import org.folio.entitlement.service.stage.CancellationFailedFlowFinalizer;
-import org.folio.entitlement.service.stage.CancelledFlowFinalizer;
-import org.folio.entitlement.service.stage.FailedFlowFinalizer;
 import org.folio.entitlement.service.stage.FinishedFlowFinalizer;
 import org.folio.entitlement.service.stage.FlowInitializer;
 import org.folio.entitlement.service.stage.TenantLoader;
@@ -35,13 +32,11 @@ public class EntitleFlowFactory implements FlowFactory {
 
   private final FinishedFlowFinalizer finishedFlowFinalizer;
   private final FlowInitializer flowInitializer;
-  private final FailedFlowFinalizer failedFlowFinalizer;
-  private final CancelledFlowFinalizer cancelledFlowFinalizer;
-  private final CancellationFailedFlowFinalizer cancellationFailedFlowFinalizer;
+  private final FlowFinalizerCallbacks finalizerCallbacks;
 
   @Override
   public Flow createFlow(EntitlementRequest request) {
-    return Flow.builder()
+    var builder = Flow.builder()
       .id(UUID.randomUUID())
       .stage(flowInitializer)
       .stage(applicationFlowValidator)
@@ -51,10 +46,9 @@ public class EntitleFlowFactory implements FlowFactory {
       .stage(applicationFlowQueuingStage)
       .stage(kafkaTenantTopicCreator)
       .stage(DynamicStage.of(applicationsFlowFactory.getName(), applicationsFlowFactory::createFlow))
-      .stage(finishedFlowFinalizer)
-      .onFlowError(failedFlowFinalizer)
-      .onFlowCancellation(cancelledFlowFinalizer)
-      .onFlowCancellationError(cancellationFailedFlowFinalizer)
+      .stage(finishedFlowFinalizer);
+
+    return finalizerCallbacks.apply(builder)
       .flowParameter(PARAM_REQUEST, request)
       .executionStrategy(request.getExecutionStrategy())
       .build();