+++ src/test/java/org/folio/entitlement/service/stage/SkippedApplicationFlowFinalizerTest.java
@@ -1,22 +1,15 @@
 package org.folio.entitlement.service.stage;
 
 import static org.folio.entitlement.domain.dto.EntitlementRequestType.ENTITLE;
-import static org.folio.entitlement.domain.entity.type.EntityExecutionStatus.IN_PROGRESS;
-import static org.folio.entitlement.domain.entity.type.EntityExecutionStatus.QUEUED;
 import static org.folio.entitlement.domain.model.ApplicationStageContext.PARAM_APPLICATION_FLOW_ID;
 import static org.folio.entitlement.domain.model.CommonStageContext.PARAM_REQUEST;
 import static org.folio.entitlement.support.TestConstants.APPLICATION_FLOW_ID;
-import static org.folio.entitlement.support.TestConstants.APPLICATION_ID;
 import static org.folio.entitlement.support.TestConstants.FLOW_STAGE_ID;
 import static org.folio.entitlement.support.TestConstants.TENANT_ID;
 import static org.folio.entitlement.support.TestValues.appStageContext;
-import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
 import java.util.Map;
-import org.folio.entitlement.domain.entity.ApplicationFlowEntity;
-import org.folio.entitlement.domain.entity.type.EntityExecutionStatus;
 import org.folio.entitlement.domain.model.EntitlementRequest;
 import org.folio.entitlement.repository.ApplicationFlowRepository;
 import org.folio.entitlement.support.TestUtils;
@@ -42,33 +35,11 @@ class SkippedApplicationFlowFinalizerTest {
 
   @Test
   void execute_positive() {
-    var entity = applicationFlowEntity(QUEUED);
-    when(applicationFlowRepository.getReferenceById(APPLICATION_FLOW_ID)).thenReturn(entity);
-
     var stageContext = appStageContext(FLOW_STAGE_ID, flowParameters(), Map.of());
+
     flowFinalizer.execute(stageContext);
 
-    verify(applicationFlowRepository).delete(entity);
-  }
-
-  @Test
-  void execute_positive_statusIsNotFinished() {
-    var entity = applicationFlowEntity(IN_PROGRESS);
-    when(applicationFlowRepository.getReferenceById(APPLICATION_FLOW_ID)).thenReturn(entity);
-
-    var stageContext = appStageContext(FLOW_STAGE_ID, flowParameters(), Map.of());
-    flowFinalizer.execute(stageContext);
-
-    verify(applicationFlowRepository, never()).delete(entity);
-  }
-
-  private static ApplicationFlowEntity applicationFlowEntity(EntityExecutionStatus status) {
-    var entity = new ApplicationFlowEntity();
-    entity.setId(APPLICATION_FLOW_ID);
-    entity.setApplicationId(APPLICATION_ID);
-    entity.setTenantId(TENANT_ID);
-    entity.setStatus(status);
-    return entity;
+    verify(applicationFlowRepository).removeQueuedFlow(APPLICATION_FLOW_ID);
   }
 
   private static Map<?, ?> flowParameters() {