+++ NEWS.md
@@ -1,4 +1,5 @@
 ## Version `v4.1.0` (In Progress)
+* Gracefully complete entitlement for desired state operation in no async mode (MGRENTITLE-161)
 * Upgrade dependencies for Kafka 4.2 compatibility in mgr-tenant-entitlements (MGRENTITLE-180)
 * Switch Kong integration test container to folioci/folio-kong image (APPPOCTOOL-37)